trained model
trained model for recognising news of ACTIVISION BLIZZARD
It is necessary to have a dataset which will have a format like
date, news; and the program will create a new column, named characteristic. In the characteristic column you will have +, - or ? for each new, where + is a positive news for the company, - is negative, and a question mark means that it is hard to say whether news will be positive or negative.